post-pylint code tidyup
